feat : DailyRecord, MonthlyRecord, WeeklyRecord 구현

+++ first/first/src/main/java/com/runtobeat/first/dto/RecordResponseDTO.java
@@ -1,6 +1,5 @@
 package com.runtobeat.first.dto;
 
-import com.fasterxml.jackson.annotation.JsonGetter;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
@@ -14,7 +13,7 @@ import java.time.LocalTime;
 public class RecordResponseDTO {
     private String recordId;
     private String memberId;
-    private Long runningDistance;
+    private Double runningDistance;
     private LocalTime runningTime;
     private Long runningStep;
     private LocalDate recordDate;
